+++ packages/subscribables/core/src/index.ts
@@ -1,9 +1,0 @@
-export {
-  type Subscribable,
-  type SubscribableObject,
-  type SubscribableValue,
-  isSubscribableValue,
-  valueOfSubscribable,
-} from "./subscribable";
-export { SubscriberSet } from "./subscriber-set";
-export { computed } from "./computed";